Moved group example back to examples
