+++ examples/group/src/main/scala/indigoexamples/GroupExample.scala
@@ -5,40 +5,54 @@ import indigo._
 import scala.scalajs.js.annotation._
 
 @JSExportTopLevel("IndigoGame")
-object GroupExample extends IndigoSandbox[Unit, Unit] {
+object GroupExample extends IndigoSandbox[Unit, Radians] {
 
-  val config: GameConfig = defaultGameConfig
+  val config: GameConfig =
+    GameConfig.default
 
-  val assetName: AssetName = AssetName("graphics")
+  val assetName: AssetName =
+    AssetName("graphics")
 
-  val assets: Set[AssetType] = Set(AssetType.Image(assetName, AssetPath("assets/graphics.png")))
+  val assets: Set[AssetType] =
+    Set(AssetType.Image(assetName, AssetPath("assets/graphics.png")))
 
-  val fonts: Set[FontInfo] = Set()
+  val fonts: Set[FontInfo] =
+    Set()
 
-  val animations: Set[Animation] = Set()
+  val animations: Set[Animation] =
+    Set()
 
   def setup(assetCollection: AssetCollection, dice: Dice): Outcome[Startup[Unit]] =
     Outcome(Startup.Success(()))
 
-  def initialModel(startupData: Unit): Outcome[Unit] =
-    Outcome(())
+  def initialModel(startupData: Unit): Outcome[Radians] =
+    Outcome(Radians.zero)
 
-  def updateModel(context: FrameContext[Unit], model: Unit): GlobalEvent => Outcome[Unit] =
-    _ => Outcome(model)
+  def updateModel(context: FrameContext[Unit], model: Radians): GlobalEvent => Outcome[Radians] = {
+    case FrameTick =>
+      Outcome(model + Radians(0.01))
+    case _ =>
+      Outcome(model)
+  }
 
-  def present(context: FrameContext[Unit], model: Unit): Outcome[SceneUpdateFragment] =
+  val graphic =
+    Graphic(0, 0, 32, 32, 1, Material.Textured(assetName))
+      .withCrop(32, 0, 32, 32)
+      .withRef(16, 16)
+
+  def present(context: FrameContext[Unit], model: Radians): Outcome[SceneUpdateFragment] =
     Outcome(
       SceneUpdateFragment.empty.addGameLayerNodes(
+        graphic.moveTo(16, 16),
         Group(
-          Graphic(0, 0, 256, 256, 1, Material.Textured(assetName)).moveTo(64, 10).moveBy(-50, -50),
-          Graphic(0, 0, 32, 32, 1, Material.Textured(assetName)).withCrop(32, 0, 32, 32).moveBy(-50, -50),
-          Graphic(0, 0, 128, 128, 1, Material.Textured(assetName))
-            .moveTo(0, 128)
-            .withCrop(128, 0, 128, 128)
-            .withTint(0, 1, 1)
-            .withAlpha(0.5)
-            .moveBy(-50, -50)
-        ).moveTo(100, 100)
+          graphic,
+          Group(graphic)
+            .moveBy(64, 64)
+            .scaleBy(2, 2)
+            .rotateBy(model * Radians(2.0))
+        ).moveBy(config.screenDimensions.center)
+          .scaleBy(2, 2)
+          .rotateBy(model)
       )
     )
 }
