+++ indigo/build.sbt
@@ -75,25 +75,6 @@ lazy val perf =
     .dependsOn(indigoExtras)
     .dependsOn(indigoJsonCirce)
 
-lazy val group =
-  project
-    .enablePlugins(ScalaJSPlugin)
-    .enablePlugins(SbtIndigo)
-    .settings(commonSettings: _*)
-    .settings(
-      name := "group",
-      showCursor := true,
-      title := "Group",
-      gameAssetsDirectory := "assets"
-    )
-    .settings(
-      publish := {},
-      publishLocal := {}
-    )
-    .dependsOn(indigo)
-    .dependsOn(indigoExtras)
-    .dependsOn(indigoJsonCirce)
-
 // Indigo
 lazy val indigoCore =
   project
@@ -229,8 +210,7 @@ lazy val indigoProject =
       indigo,
       indigoFacades,
       sandbox,
-      perf,
-      group
+      perf
     )
 
 lazy val code =
