+++ indigo/commands.sbt
@@ -12,8 +12,7 @@ lazy val releaseProjects: List[String] =
 lazy val coreProjects: List[String] =
   releaseProjects ++ List(
     "sandbox",
-    "perf",
-    "group"
+    "perf"
   )
 
 def applyCommand(projects: List[String], command: String): String =
@@ -237,12 +236,4 @@ addCommandAlias(
 addCommandAlias(
   "crossIndigoPublishAllSigned",
   applyCrossToAllReleaseable("publishSigned")
-)
-addCommandAlias(
-  "groupRun",
-  List(
-    "buildAllNoClean",
-    "group/fastOptJS",
-    "group/indigoRun"
-  ).mkString(";", ";", "")
 )
